+++ changelog.md
@@ -6,7 +6,6 @@
 * Torch placer (Chestplate)
 * Wither effect on hit (treasure)
 * Celestial Guardian (gives absorption when below a certain health) (shield)
-* Smelting milk should give back bottle
 * Smartness should work with fishing rods too
 * Repairing an enchanted item with an unenchanted costs levels
 * Enchanting in creative consumes levels
@@ -15,6 +14,7 @@
 
 ## Upcoming
 * Blast furnaces now try to pick up an item every 5 seconds if empty
+* Smartness now works on Fishing Rods
 * Fixed enchanting table learning curses
 
 ## 1.6.1